refactor(protocol-v2): widen claim mask to u64 (design doc §15 R2)
The resource-claim mask was u32 (axes 0-7, DACs 8-15, named 16-21), which caps
DACs at 8 and is already full on the current axis list. Widen to u64 and
re-layout: axes 0-15, DACs 16-31, named resources 32+. New PCBs (e.g. 16 DACs)
now need no re-layout. The claim mask is internal — never on the wire — so this
does NOT touch the frozen 158-byte StandardResponse; golden vectors are
unchanged. (Wire response array sizing = R3, deliberately deferred.)

frames.h res_axis/res_dac/RES_* and the Python mirror move in lockstep (the
frames.h parser test enforces it); claims/slots/dispatch signatures and the
conflict-scan loop go u64; test expectations updated (HEX64, RES_ILLUM_TTL now
bit 32 -> conflict index 33).

Verified: pio test -e native (147), pytest (19), pio run -e teensy41 and
-e teensy41_boardv2, cppcheck clean, golden generator idempotent.

+++ firmware/controller/src/protocol/claims.h
@@ -17,16 +17,16 @@
 namespace protocol {
 
 // Resource mask an incoming command claims, from the production table.
-uint32_t claims_for(uint8_t cmd_type, const uint8_t* payload, size_t len);
+uint64_t claims_for(uint8_t cmd_type, const uint8_t* payload, size_t len);
 
 // Same lookup against an explicit table (tests and Phase D supply their own).
-uint32_t claims_for_in(const ClaimsRow* table, size_t count, uint8_t cmd_type,
+uint64_t claims_for_in(const ClaimsRow* table, size_t count, uint8_t cmd_type,
                        const uint8_t* payload, size_t len);
 
 // 0 if `wanted` and `inflight_union` share no resources; otherwise the
 // lowest-set conflicting resource bit index + 1 (so 0 unambiguously means
 // "no conflict", and the caller recovers the resource id by subtracting 1).
-uint8_t claims_conflict(uint32_t wanted, uint32_t inflight_union);
+uint8_t claims_conflict(uint64_t wanted, uint64_t inflight_union);
 
 }  // namespace protocol
